REFAC: move widgets to external classes
